+++ apps/web-client/src/core/query/episode-progress.ts
@@ -10,7 +10,7 @@ import {
   type SeasonProgressDto,
 } from '../api/episode-progress.client';
 import { queryKeys } from './keys';
-import { isUnauthorized } from './utils';
+import { retryUnlessUnauthorized } from './utils';
 
 /**
  * Invalidates all caches affected by episode progress changes.
@@ -36,13 +36,6 @@ function invalidateEpisodeProgressCaches(queryClient: QueryClient, showId: strin
 // Query Hooks
 // ============================================================================
 
-/**
- * Fetches watch progress for all seasons of a show.
- *
- * @param showId - Show UUID
- * @param options - Additional query options
- * @returns Query result with season progress
- */
 export function useShowProgress(
   showId: string | undefined,
   options?: Omit<UseQueryOptions<ShowProgressDto>, 'queryKey' | 'queryFn'>,
@@ -52,10 +45,7 @@ export function useShowProgress(
     queryFn: () => episodeProgressApi.getShowProgress(showId!),
     enabled: !!showId,
     staleTime: 1000 * 60 * 5, // 5 minutes
-    retry: (failureCount, error) => {
-      if (isUnauthorized(error)) return false;
-      return failureCount < 2;
-    },
+    retry: retryUnlessUnauthorized,
     ...options,
   });
 }
@@ -80,12 +70,6 @@ interface MarkAllEpisodesVariables {
   episodesBySeasonNumber: Map<number, string[]>;
 }
 
-/**
- * Toggles episode watched status with optimistic updates.
- *
- * @param showId - Show UUID for cache invalidation
- * @returns Mutation with toggle function
- */
 export function useToggleEpisodeWatched(showId: string) {
   const queryClient = useQueryClient();
 
@@ -155,13 +139,6 @@ export function useToggleEpisodeWatched(showId: string) {
   });
 }
 
-/**
- * Marks multiple episodes as watched with optimistic updates.
- * Used for "mark previous episodes" feature.
- *
- * @param showId - Show UUID for cache invalidation
- * @returns Mutation with bulk mark function
- */
 export function useMarkMultipleWatched(showId: string) {
   const queryClient = useQueryClient();
 
@@ -224,10 +201,6 @@ export function useMarkMultipleWatched(showId: string) {
   });
 }
 
-/**
- * Marks ALL episodes of a show as watched.
- * Returns previous state for undo functionality.
- */
 export function useMarkAllEpisodesWatched(showId: string) {
   const queryClient = useQueryClient();
 
@@ -285,9 +258,6 @@ export function useMarkAllEpisodesWatched(showId: string) {
   });
 }
 
-/**
- * Unmarks episodes (for undo). Accepts episode IDs directly to avoid race conditions.
- */
 export function useUnmarkEpisodes(showId: string) {
   const queryClient = useQueryClient();
