+++ apps/web-client/src/core/query/utils.ts
@@ -1,5 +1,10 @@
 import { HTTPError } from 'ky';
 
-export function isUnauthorized(error: unknown): boolean {
+function isUnauthorized(error: unknown): boolean {
   return error instanceof HTTPError && error.response.status === 401;
+}
+
+export function retryUnlessUnauthorized(failureCount: number, error: unknown): boolean {
+  if (isUnauthorized(error)) return false;
+  return failureCount < 2;
 }
